requirements.txt
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
cryptography==44.0.2
enum-compat==0.0.3
idna==3.10
intuit-oauth==1.2.6
oauthlib==3.2.2
pycparser==2.22
PyJWT==2.10.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-quickbooks==0.9.11
requests==2.32.3
requests-oauthlib==2.0.0
setuptools==78.1.0
six==1.17.0
urllib3==2.4.0
